+++ api/middleware.go
@@ -1,0 +1,53 @@
+package api
+
+import (
+	"errors"
+	"github.com/gin-gonic/gin"
+	"net/http"
+	"strings"
+	"whaleWake/token"
+)
+
+const (
+	authorizationHeaderKey  = "authorization"
+	authorizationTypeBearer = "bearer"
+	authorizationPayloadKey = "authorization_payload"
+)
+
+func authMiddleware(tokenMaker token.Maker) gin.HandlerFunc {
+	return func(ctx *gin.Context) {
+		authorizationHeader := ctx.GetHeader(authorizationHeaderKey)
+
+		if len(authorizationHeader) == 0 {
+			err := errors.New("authorization header is empty")
+			ctx.AbortWithStatusJSON(http.StatusUnauthorized, err)
+			return
+		}
+
+		// Extract the token from the header
+		fields := strings.Fields(authorizationHeader)
+		if len(fields) < 2 {
+			err := errors.New("invalid authorization header format")
+			ctx.AbortWithStatusJSON(http.StatusUnauthorized, err)
+			return
+		}
+		authorizationType := strings.ToLower(fields[0])
+		if authorizationType != authorizationTypeBearer {
+			err := errors.New("unsupported authorization type")
+			ctx.AbortWithStatusJSON(http.StatusUnauthorized, err)
+			return
+		}
+
+		accessToken := fields[1]
+		payload, err := tokenMaker.VerifyToken(accessToken)
+		if err != nil {
+			ctx.AbortWithStatusJSON(http.StatusUnauthorized, err)
+			return
+		}
+
+		// Store the payload in the context for later use
+		ctx.Set(authorizationPayloadKey, payload)
+		ctx.Next() // Proceed to the next handler
+
+	}
+}